Added Policy MVAC Modeling
Modeled MVAC consumption, banks, and emissions based on the EU proposed
policy targets for consumption and the percentage breakdowns for MVAC’s
portion of consumption. Extrapolated MVAC consumption to 2034, and then
set 2035-2050 consumption constant since that was the pattern of the
total consumption in the policy scenario. Linear extrapolation by
itself led to negative consumption values for MVAC. I was considering
using linear extrapolation to model the changing MVAC percentages (of
total consumption) and then applying those percentages to the total
consumption policy goals.