Plotting
Added two sets of looping to generate plots. One set generates a plot
for each variable and for each region (A5 and Non-A5 countries), with
each plot containing all of the scenarios for that vector.
The other set of looping generates a plot of one scenario and one
region, plotting at least two variables on the same axes.